rename: components/navbar/NavBar로 파일 이동
