+++ src/layouts/DefaultLayout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from "react-router-dom";
 import ScrollToTop from "../components/ScrollToTop";
-import NavBar from "../components/NavBar";
+import NavBar from "../components/navbar/NavBar";
 import Footer from "../components/Footer";
 
 const DefaultLayout = () => {
